refactor(loop): split php
