+++ packages/wp-plugin/ionos-essentials/ionos-essentials/inc/loop/schema.json
@@ -37,12 +37,7 @@
         "locale": {
           "type": "string",
           "description": "WordPress locale setting",
-          "examples": [
-            "fr_FR",
-            "en_US",
-            "de_DE",
-            "de_DE_formal"
-          ]
+          "examples": ["fr_FR", "en_US", "de_DE", "de_DE_formal"]
         },
         "blog_public": {
           "type": "boolean",
@@ -51,35 +46,22 @@
         "market": {
           "type": "string",
           "description": "Tenant market identifier",
-          "examples": [
-            "FR",
-            "US",
-            "DE"
-          ]
+          "examples": ["FR", "US", "DE"]
         },
         "tenant": {
           "type": "string",
           "description": "Tenant identifier",
-          "examples": [
-            "ionos"
-          ]
+          "examples": ["ionos"]
         },
         "core_version": {
           "type": "string",
           "description": "WordPress core version",
-          "examples": [
-            "6.8.1",
-            "5.9.3"
-          ]
+          "examples": ["6.8.1", "5.9.3"]
         },
         "php_version": {
           "type": "string",
           "description": "PHP version",
-          "examples": [
-            "8.1",
-            "8.2",
-            "7.4"
-          ]
+          "examples": ["8.1", "8.2", "7.4"]
         },
         "installed_themes": {
           "type": "integer",
@@ -94,19 +76,13 @@
           "description": "UTC Unix timestamp when the instance was created"
         },
         "last_login": {
-          "type": [
-            "integer",
-            "null"
-          ],
+          "type": ["integer", "null"],
           "description": "UTC Unix timestamp of the last login"
         },
         "permalink_structure": {
           "type": "string",
           "description": "WordPress permalink structure",
-          "examples": [
-            "/%postname%/",
-            "/%year%/%monthnum%/%postname%/"
-          ]
+          "examples": ["/%postname%/", "/%year%/%monthnum%/%postname%/"]
         },
         "siteurl": {
           "type": "string",
@@ -129,11 +105,7 @@
       "description": "Currently active theme information",
       "items": {
         "type": "object",
-        "required": [
-          "id",
-          "version",
-          "auto_update"
-        ],
+        "required": ["id", "version", "auto_update"],
         "properties": {
           "id": {
             "type": "string",
@@ -144,10 +116,7 @@
             "description": "Theme version"
           },
           "parent_theme_slug": {
-            "type": [
-              "string",
-              "null"
-            ],
+            "type": ["string", "null"],
             "description": "Parent theme slug for child themes"
           },
           "auto_update": {
@@ -163,11 +132,7 @@
       "description": "List of active plugins",
       "items": {
         "type": "object",
-        "required": [
-          "plugin_slug",
-          "version",
-          "auto_update"
-        ],
+        "required": ["plugin_slug", "version", "auto_update"],
         "properties": {
           "plugin_slug": {
             "type": "string",
@@ -195,18 +160,12 @@
       "description": "Post type statistics",
       "items": {
         "type": "object",
-        "required": [
-          "type",
-          "count"
-        ],
+        "required": ["type", "count"],
         "properties": {
           "type": {
             "type": "string",
             "description": "Post type",
-            "examples": [
-              "post",
-              "page"
-            ]
+            "examples": ["post", "page"]
           },
           "count": {
             "type": "integer",
@@ -219,10 +178,7 @@
     "comments": {
       "type": "object",
       "description": "Comment statistics and settings",
-      "required": [
-        "total",
-        "comments_active"
-      ],
+      "required": ["total", "comments_active"],
       "properties": {
         "total": {
           "type": "integer",
@@ -240,32 +196,16 @@
       "description": "Timeline of events",
       "items": {
         "type": "object",
-        "required": [
-          "name",
-          "payload",
-          "timestamp"
-        ],
+        "required": ["name", "payload", "timestamp"],
         "properties": {
           "name": {
             "type": "string",
             "description": "Event type",
-            "examples": [
-              "login",
-              "plugin install",
-              "theme change",
-              "post published"
-            ]
+            "examples": ["login", "plugin install", "theme change", "post published"]
           },
           "payload": {
             "description": "Event-specific payload data",
-            "type": [
-              "object",
-              "array",
-              "string",
-              "number",
-              "boolean",
-              "null"
-            ],
+            "type": ["object", "array", "string", "number", "boolean", "null"],
             "default": null
           },
           "timestamp": {
@@ -279,10 +219,7 @@
     "uploads": {
       "type": "object",
       "description": "Media library statistics",
-      "required": [
-        "file_count",
-        "file_size"
-      ],
+      "required": ["file_count", "file_size"],
       "properties": {
         "file_count": {
           "type": "integer",
